Pulled scrape methods into own module for vieshow
